requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
elasticsearch==7.15.2
fastapi==0.70.0
frozenlist==1.2.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
idna==3.3
multidict==5.2.0
orjson==3.6.4
pycodestyle==2.8.0
pydantic==1.8.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
twilio==7.3.1
typing_extensions==4.0.0
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
